First cut at removing buffering. This update removes it on the read side only. Warning: I've quite possibly broken some things, but there is now just one reading macro that is instantiated for all the various sizes, so it should be easy to fix.
Tested for some basic conversions, and ALSA playback (which needed
some work as it relied on the old behaviour of the buffering
functions, so I've inlined them in alsa.c instead).

The conversion macros are now a bit more regular: many more take a
clip count argument, even when they don't use it.