Oh dear. I left sox unable to produce output, as cleanup would always remove the output file. This is fixed.
However, it turned out that flowing was reading the -1th effect, which
isn't a good idea (and I thought that was the cause of the output
problem for a long time, duh).

I made a lot more cosmetic fixes to the source, in particular fixing
up comments.

Finally, I removed some of the "intelligent" code: running all effects
after avg when reducing channels, and running all effects after
downsampling. This is because this may not be what you actually want,
and the speed gain (or loss) won't typically be great, and is in any
case entirely predictable (assuming most of the algorithms are linear
or log-linear, it'll be roughly proportional to the ratio of output
data to input data), so it seemed better to remove this source of bugs
and user confusion at the expense of a little more intelligence on the
user's part required for high-performance conversion or
low-performance computers.