Add a --disable flag for each sound driver.
Change the sound drivers to "--disable/enable" rather than
"--with-without" to match the suggested semantics of each
(disable/enable being for built-in code, and with-without for add-on
extras).

Overhaul the use of enable/disable in general to make the code more
uniform and, in the case of silencing libtool, correct, basing it on
the debug flag code, which was the best.

Add code to detect -lossaudio (assume that if it exists it's needed).

In sum, these changes close #1666597.