Fix crash in compand effect
The compand effect drain() function overran its output buffer if the
buffer length was not evenly divisible by the number of channels (i.e.
usually, if the number of channels was not a power of two). This was
only triggered if a delay was specified, since otherwise there are no
samples to be drained.