Allow the last of multiple effects chains to be empty
In multiple effects chain mode, allow the last chain to be empty to
indicate copying without further processing. E.g.,
  sox infile.wav outfile.wav trim 0 30 reverse :
would reverse the first thirty seconds of the input file, then append
the rest of the file unchanged.

This resolves feature request 2459231 and also fixes a memory leak,
where in previous versions the empty last chain was ignored but not
freed.