Rewritten trim effect
From-scratch rewrite of the trim effect. Any number of parameters may
be given, and the effect will alternate between discarding and copying
audio at each position. Time values may also be specified relative to
the end of audio, if audio length is known.

The old "trim" syntax is a special case of the extended syntax (only
one or two parameters allowed, first must not use "=", no positions
relative to end-of-audio), so backwards compatibility is maintained.

This also resolves feature request 2941349.