Fix optimize_trim() in multiple effects chain mode
Try the optimize_trim() hack only at the very start of processing, but
not after switching to another effects chain or restarting the first.
It would otherwise seek relative to the beginning of the input file,
changing the semantics of the command instead of just providing a
speed improvement. To make doubly sure and developers aware of the
trap, check for the condition both within optimize_trim() and in
process() where it is called.

To illustrate the bug fixed by this commit, consider the command line
  play example trim 1s 10 : trim 1s 10
which is now (correctly) almost equivalent to
  play example trim 0 20
but was similar to
  play example trim 0 10 repeat
before, at least in case "example" was a seekable file.

The problem was concealed in many cases (e.g., the command line
  sox infile.wav output.wav trim 0 30 : newfile : restart
given as an example in the man page) by the fact that optimize_trim()
didn't kick in when the offset was zero.