Don't drop samples at effects chain transitions
When using multiple/restarted effects chains, at each transition to
or restart of an effects chain, a number of samples would be lost,
even if the effect causing the transition was the first in its chain
(as recommended in the man page). The exact number of affected
samples would depend on buffer size, number of channels, flow
behaviour of the effects/handlers involved etc. and was thus
unpredictable.

To correct this problem, this commit modifies sox_flow_effects() to
not free effects' output buffers at the end of processing (this is
left to sox_delete_effect() instead). Since the input combiner
effect is reused in the new effects chain, this allows samples
already processed by the combiner to remain in its buffer and be
flowed down the new effects chain later.

This should, a.o., make sure that after
  sox infile splitfiles trim 0 1:00 : newfile : restart
  sox splitfiles* recombined
"infile" and "recombined" are identical (provided the encoding is
lossless, of course).

(Note that samples buffered by "normal" effects, i.e. those that are
not reused, to the left of the effect terminating the chain will still
be dropped. E.g., with
  sox infile outfile highpass 40 trim 0 10 :
a number of samples already processed by the highpass effect but not
consumed by trim will be lost. This behaviour is already documented in
the man page.)