Reorder effect and effects chain structure members
Put members of sox_effect_t and sox_effects_chain_t that are for
internal use only to the end of the structures.

This breaks ABI compatibility, but will make it easier to change them
in future if needed, since offsets for non-private members won't have
to change again then.