Check if pulseaudio is actually available
This is a partial backport of robs' commit 55c0aa7 (check that writable
devices can actually be openned for writing, before selecting as the
default device), limited to pulseaudio where it seems to work well.
If the pulseaudio format is compiled in, but the PulseAudio server is not
actually usable, SoX will fall through to the next available format.
Debian bug #676167 shows that the full solution throws a warning with
alsa and even segfaults if it reaches the oss driver.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=676167