Read Ogg Vorbis files with any number of channels
Reading an Ogg Vorbis file with 3, 5, 6, etc. (any non-power of two)
channels terminated early after a few hundred samples (Debian
bug #672567).

In this situation, there always remained a few unused bytes in the read
buffer, as its size (4096 bytes) wasn't evenly divisible by the number of
channels. Still, ov_read() got called with this practically exhausted
buffer and failed.

To solve this, the buffer is created so that it always can hold an integer
number of wide samples, deviating slightly from the default size if
necessary.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=672567