delay: allow mixing time and sample-count args
Parse the arguments to the delay effect when the sample rate is known
(i.e. in start()). Previously, a dummy rate (100000) was used to do the
parsing on creation, which could lead to the maximum delay not being
identified correctly. (Example: sample rate = 48000, arguments "1" and
"50000s"; here, the first argument was erroneously identified as the
maximum, being interpreted as 100000 samples even though it actually
only represented 48000.)

Update the man page to no longer warn against mixing the two methods
of specifying delay values.