Improve lsx_parsesamples syntax checking
Make lsx_parsesamples more predictable regarding which characters it
considers to belong to a time specification. Any sequence matching the
regex ^[0-9:.est]* is now considered the scope of lsx_parsesamples and
consumed completely. It is an error if characters out of [0-9:.est]
remain after processing. Other characters (like '@' or ',') are still
OK.

In particular, this means that input strings like "1.2e3t" and "1ts"
are now illegal. Previously, the former was interpreted as 1+0.2e3 =
1+200 = 201 seconds instead of 1200 seconds as a user reasonably might
have expected; the latter was interpreted as 1 sample due to the 's'
at the end, yet only "1" was consumed and "ts" returned for further
processing by the caller.