Extend time specification syntax
Allow time specifications to be chained, adding/subtracting the specified
amounts; e.g.,
  trim =2:00-100s =2:00+100s
will cut 200 samples centered around 2:00. This is especially useful when
combining time-based and sample-based specifications, as well as for
scripts that are given audio positions as parameters and need to add an
offset to those.

It is also possible for such a chained time specification to comprise
more than two parts, as in `2:00-12+13s-1:23'. No immediate use case
presents itself for that, but the code cost to implement this feature is
negligible.