Introduce an extended syntax for audio positions
Introduce a new lsx_parseposition function (similar to lsx_parsesamples)
that makes a trim-like syntax for specifying audio positions available to
all effects; it is possible to give a position relative to the start of
audio ("=2:00"), to the end of audio ("-1:30"), or to a previous position
("+1:00").

The original lsx_parsesamples function is replaced by a wrapper for an
internal parsesamples function that allows a start value and an initial
combine mode (add/subtract) to be passed in, which lsx_parseposition
needs.

This commit only introduces the function and common documentation. No
effects are using lsx_parseposition yet.