splice: set output length to unknown, not 0
In the splice effect, the output length was only set (to unknown) when it
was not needed, immediately before returning with SOX_EFF_NULL. This
resulted in an incorrect output length of 0 when the effect had work to
do. Move the relevant line of code to a more appropriate place.