Apply new timespec syntax to documentation example
Update the "acpo" example script shown in the man page to take
advantage of the syntax features that were added recently (chained
time specifications, positions). All calculations that previously had
to be done in the shell can now be done in SoX directly, and values
like "1:03" instead of only sample numbers are allowed as parameters
to the script.