Don't export (most) internal libsox symbols
Ideally, libsox should only export those symbols that are part of the
official API, i.e. those with the sox_ prefix. This isn't easily
possible, though, because certain internal (lsx_) symbols are used by
SoX-the-application and by the plugin modules in case of a modularized
build. As a step in the right direction, this change exposes only the
symbols actually needed, hiding about 300 (90 %) of the lsx_ symbols.

(Relevant tracker item: FR 2214694.)