ladspa: optionally perform latency compensation
Some LADSPA plugins report their latency using a output control
port named "latency".  We may capture the latency of the plugin
and from that port to compensate for latency the plugin introduced.

This feature exists in Ardour and appears enabled by default[1].
This is not enabled by default in sox as existing scripts for older
versions of sox may already compensate for latency.

I've verified the output with the following no-op commands using on
tap_limiter (0.7.2) and sndfile-cmp on the results using the
following script:

----------------------------- 8< ----------------------------
  set -e

  # a.flac is a 24-bit, stereo, 44.1kHz FLAC file
  sox -n -r 48000 -c2 -b24 a.flac synth 5 pluck

  # original 24-bit FLAC, convert to 32-bit float and back to 24-bit wav
  sox a.flac -tf32 -c2 -r48000 - | sox -tf32 -r48000 -c2 - -b24 a.wav

  # no-op operation for tap_limiter, LADSPA uses float internally so
  # we expect this to give the same output as the above sox-only pipeline
  sox a.flac b.wav ladspa -rl tap_limiter 0 0

  sndfile-cmp a.wav b.wav
----------------------------- 8< ----------------------------

[1] - Based on my (limited) understanding of the Ardour code.
      I have not tested with Ardour as I am not comfortable with GUIs.