stats: improve bit depth reporting [bug #267]
The right-hand figure reported for bit depth is somewhat erratic,
often changing if the signal is inverted or changed in other ways
that should not affect the value.

Change the calculation such that it always reports the number of
bits required to represent the largest sample value, positive or
negative, in two's complement notation.