wav: allow valid bits less than sample size
Nothing special action is required for files with number of valid
bits less than the sample size.  The reported precision will be
incorrect, but that is less bad than refusing to process the file
at all.

If the specified valid bits is greater than the sample size, the
file is rejected as per Microsoft recommendation.