build: simplify win32 glob selection
Move the files to a subdirectory so the header can be activated by
simply adding a -I flag, avoiding ugly ifdefs.  Use this version if
a native glob is missing and FindFirstFileA() exists.  This is more
accurate than matching the system name since some mingw versions do
have a glob implementation.